Rename subgraph.rs to parser.rs. Move and rename some functions and tests from config.rs
